requirements.txt
asgiref==3.4.1
autopep8==1.5.7
Django==3.2.8
django-filter==21.1
djangorestframework==3.12.4
numpy==1.21.2
opencv-python==4.5.3.56
Pillow==8.3.2
pycodestyle==2.7.0
python-barcode==0.13.1
pytz==2021.3
qrcode==7.3.1
sqlparse==0.4.2
toml==0.10.2
